+++ src/encoded/static/components/index.js
@@ -30,6 +30,7 @@ import SearchView from './browse/SearchView';
 import BrowseView from './browse/BrowseView';
 import FileSearchView from './browse/FileSearchView';
 import FileView from './item-pages/FileView';
+import PublicDonorView from './item-pages/PublicDonorView';
 import ProtectedDonorView from './item-pages/ProtectedDonorView';
 
 /**
@@ -52,6 +53,7 @@ content_views.register(UserView, 'User');
 content_views.register(ImpersonateUserForm, 'User', 'impersonate-user');
 content_views.register(StaticSectionView, 'StaticSection');
 content_views.register(FileView, 'File');
+content_views.register(PublicDonorView, 'Donor');
 content_views.register(ProtectedDonorView, 'ProtectedDonor');
 
 content_views.register(SMaHTSubmissionView, 'Item', 'edit');